refactor: modify folder name dto to dtos
