+++ webapp-test/server/src/app/user/user.service.ts
@@ -13,7 +13,7 @@ import {
   AllUserResponseDto,
   ToggleFollowResponseDto,
   UserInfoResponseDto,
-} from "./dto/user.dto";
+} from "./dtos/user.dto";
 
 @Injectable()
 export class UserService {
